refactor(deliberation): update import paths for components
Reorganized import paths within the deliberation module to simplify and improve maintainability. This change replaces relative paths pointing to `page.client` with direct references to the parent directory. Additionally, added a new `index.tsx` file to consolidate and streamline the deliberation module structure.
