+++ src/app/spaces/[id]/page.tsx
@@ -2,7 +2,7 @@ import { SpaceType } from '@/lib/api/models/spaces';
 import { getSpaceById } from '@/lib/api/ratel_api.server';
 import { logger } from '@/lib/logger';
 import React from 'react';
-import DeliberationSpacePage from './deliberation/page.client';
+import DeliberationSpacePage from './deliberation';
 import CommitteeSpacePage from './committee/page.client';
 
 export default async function Page({
